Deduplicate successful tasks.
This feature allows the user to deduplicate successful tasks which acks late.

The trace function fetches the metadata from the backend each time it receives a task and compares its state.
If the state is SUCCESS we log and bail instead of executing the task.
The task is acknowledged and everything proceeds normally.